Injecting Modules properly implemented:
Modules are found and instantiated at the start of the program. A dict is created with <string name, IModule instance>.

This dict is then passed into the Environment when created, becoming an Env 'Modules'.

When import statements are interpreted, they use 'GetModule' with the import name to find the IModlue from the Modules dict, then store that to the Env's variables dict as <string alias, IModule module>.

This means that while all modules are loaded at runtime, only those that the game author ha imported will bcome accessable / 'enter the gamestate', giving proper control over what ends up in the namespace and how.

Cleanup:
Cleaned 'using's across all files.

+++ CYA Adventure Game Engine/DSL/AST/Expression/PrefixExpr.cs
@@ -1,9 +1,4 @@
 ﻿using CYA_Adventure_Game_Engine.DSL.Frontend;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Environment = CYA_Adventure_Game_Engine.DSL.Runtime.Environment;
 namespace CYA_Adventure_Game_Engine.DSL.AST.Expression
 {
@@ -34,13 +29,13 @@ namespace CYA_Adventure_Game_Engine.DSL.AST.Expression
                 case TokenType.Plus:
                     return operand;
                 case TokenType.Minus:
-                    if (operand is not double)
+                    if (operand is not float)
                     {
                         throw new Exception($"Invalid value taking '-' prefix of type: {operand.GetType()}");
                     }
                     else
                     {
-                        return -(double)operand;
+                        return -(float)operand;
                     }
                 case TokenType.Not:
                     if (operand is not bool)
